fix: filter clear_notification to mobile only; arm door dwell on restore; ruff
- manager._clear_clean_notification: restrict services to mobile_app targets
  only before sending the 'clear_notification' marker; non-mobile platforms
  (email, Telegram) would otherwise receive the literal text as a message
- manager.async_setup_entry restoration: arm the door-end dwell timer when
  restoring a STATE_ENDING cycle with the door already open on auto-open
  devices — _on_state_change is not called during snapshot restoration, so
  the ENDING+open-door guard added in the prior commit would not fire
- profile_store._resample_trace_to_grid: remove redundant int() wrapping
  around round() (Ruff RUF046; round() with one arg returns int in Python 3)

+++ custom_components/ha_washdata/profile_store.py
@@ -4613,7 +4613,7 @@ class ProfileStore:
         t_max = min(float(off_arr[-1]), 2.0 * span) if span > 0 else float(off_arr[-1])
         if step <= 0 or t_max <= 0:
             return list(powers)
-        n = max(2, int(round(t_max / step)) + 1)
+        n = max(2, round(t_max / step) + 1)
         grid = np.arange(n, dtype=float) * step
         return np.interp(grid, off_arr, pow_arr).tolist()
